backend/requirements.txt
Django==4.2.14
psycopg2-binary==2.9.9

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0

channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
drf-spectacular==0.27.2

celery==5.4.0
redis==5.0.7

stripe==9.12.0

## Sentry (опционально, включается по SENTRY_DSN)
sentry-sdk>=2,<3

django-environ==0.11.2
Pillow==10.4.0

# Dev / тесты
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0

# Кодстайл и типы (dev)
flake8==7.1.0
black==24.4.2
isort==5.13.2
mypy==1.10.1
